cgroup: move setting resources to new file
