+++ src/libcrun/cgroup-resources.h
@@ -1,0 +1,29 @@
+/*
+ * crun - OCI runtime written in C
+ *
+ * Copyright (C) 2017, 2018, 2019 Giuseppe Scrivano <giuseppe@scrivano.org>
+ * crun is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation; either version 2.1 of the License, or
+ * (at your option) any later version.
+ *
+ * crun is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with crun.  If not, see <http://www.gnu.org/licenses/>.
+ */
+#ifndef CGROUP_RESOURCES_H
+#define CGROUP_RESOURCES_H
+
+#include "container.h"
+#include "cgroup.h"
+#include <unistd.h>
+
+int update_cgroup_resources (const char *path,
+                             runtime_spec_schema_config_linux_resources *resources,
+                             libcrun_error_t *err);
+
+#endif
